refactor list items to be more reusable and manageable
