+++ lib/widgets/common/list_container.dart
@@ -1,0 +1,28 @@
+import 'package:flutter/material.dart';
+
+/// Wraps some fancy decorations around a list of widgets.
+class ListContainer extends StatelessWidget {
+  const ListContainer({super.key, required this.children});
+
+  final List<Widget> children;
+
+  @override
+  Widget build(BuildContext context) {
+    return Card(
+      clipBehavior: Clip.hardEdge,
+      elevation: 0,
+      color: Theme.of(context).colorScheme.primaryContainer,
+      borderOnForeground: true,
+      shape: RoundedRectangleBorder(
+        side: BorderSide(
+          width: 4,
+          color: Theme.of(context).colorScheme.primaryContainer,
+        ),
+        borderRadius: const BorderRadius.all(Radius.circular(4)),
+      ),
+      child: Column(
+        children: children,
+      ),
+    );
+  }
+}
